+++ src/components/SearchField/index.vue
@@ -9,6 +9,7 @@
       type="text"
       placeholder="Search FAIR Data Point..."
     >
+    <small><router-link to="/search">Advanced</router-link></small>
   </form>
 </template>
 <script lang="ts">
@@ -19,9 +20,11 @@ import { Component, Vue, Watch } from 'vue-property-decorator'
 export default class SearchField extends Vue {
   q: string = ''
 
+  routeQ: string = null
+
   submit() {
     if (this.q.length > 0) {
-      this.$router.push(`/search?q=${this.q}`)
+      this.$router.push({ path: '/search', query: { q: this.q } }).catch(() => {})
     }
   }
 
@@ -31,7 +34,8 @@ export default class SearchField extends Vue {
 
   @Watch('$route')
   setQ() {
-    this.q = _.get(this.$route, 'query.q', '')
+    this.routeQ = _.get(this.$route, 'query.q', '')
+    this.q = this.routeQ
   }
 }
 </script>